requirements.txt
pyvespa
pillow
pdf2image
pypdf
openai
numpy
python-dotenv
torch
infinity_emb
